requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
flask-cors==3.0.10
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
onnxruntime==1.15.1
python-dotenv==1.0.0